RF {deepSNV}	R Documentation

Relative frequencies.

Description

Convenience function to compute the relative frequencies from a matrix with absolute counts.

Usage

  RF(freq, total = FALSE)

Arguments

freq
A matrix with nucleotide counts.

total
If the nucleotide counts have columns for forward and reverse direction, return each strand sepratatelu (FALSE), or add the two (TRUE).

Value

A matrix with the relative frequencies.

Author(s)

Moritz Gerstung

Examples

data(HIVmix)
RF(test(HIVmix))[1:10,]
RF(test(HIVmix), total=TRUE)[1:10,]

[Package deepSNV version 1.10.0 Index]