control {deepSNV}	R Documentation

Get control counts

Description

Convenience function to obtain the control counts from a deepSNV object.

Arguments

deepSNV
a deepSNV-class object

total
Logical. If true the sum of both strands is returned

Value

A matrix with the absolute frequencies summed over both strands.

Examples

data(HIVmix)
control(HIVmix)[1:10,]
control(HIVmix, total=TRUE)[1:10,]

[Package deepSNV version 1.10.0 Index]