estimateDirichlet {deepSNV}	R Documentation

Learn a base-specific Dirichlet prior.

Description

The prior learns the parameters of a Dirichlet distribution seperately for each consensus base. The expected value of the Dirichlet distributions is the base-substitution matrix, where rows correspond to the initial nucleotide and columns to the substituted nucleotide. The absolute values determine the higher moments of the Dirichlet distributions. After having learned the prior the deepSNV-class test is recomputed.

Arguments

control
Either a matrix with nucleotide counts or a deepSNV-class object.

Value

An deepSNV-class object.

Author(s)

Moritz Gerstung

Examples

data(phiX)
estimateDirichlet(phiX)

[Package deepSNV version 1.10.0 Index]