estimateRho {deepSNV}	R Documentation

Helper function for estimating the dispersion factor rho

Description

It uses a method of moments approximation to estimate rho from the variances of the relative frequencies nu across samples

Usage

  estimateRho(x, mu, ix, pseudo.rho = .Machine$double.eps)

Arguments

x
counts

mu
relative frequency across all samples

ix
index indicating the set of samples to use (typically indicating those with relative frequency smaller than 0.1).

pseudo.rho
a pseudo count added to each sample to avoid problems with zeros. Default = .Machine$double.eps

Value

rho

Note

Experimental code, subject to changes

Author(s)

mg14

[Package deepSNV version 1.10.0 Index]