loadAllData {deepSNV}	R Documentation

Function to load all data from a list of bam files

Description

This function uses the parallel package and the bam2R interface to load all nucleotide counts from a list of bam files and a set of regions into a large array.

Usage

  loadAllData(files, regions, ..., mc.cores = 1)

Arguments

files
A character vector with the paths to all bam files

regions
Either a GRanges or data.frame with the coordinates of interest

...
Arguments passed to bam2R

mc.cores
Number of cores used for loading, default = 1

Value

counts

Note

Experimental code, subject to changes

Author(s)

mg14

[Package deepSNV version 1.10.0 Index]