makePrior {deepSNV}	R Documentation

Compute a prior from a COSMIC VCF object

Description

This function computes the prior probability of detecting a true variant from a variation data base. It assumes a VCF file with a CNT slot for the count of a given base substitution. Such a VCF file can be downloaded at ftp://ngs.sanger.ac.uk/production/cosmic/. The prior probability is simply defined as pi.mut * CNT[i]/sum(CNT). On sites with no count, a background probability of pi0 is used.

Usage

  makePrior(COSMIC, regions, pi.gene = 0.1,
    pi.backgr = 1e-04)

Arguments

COSMIC
A VCF object from COSMIC VCF export.

regions
A GRanges object with the regions (gene) of interest.

pi.gene
Probability that a gene is mutated

pi.backgr
Background probability of a locus being mutated. Default 1e-4, corresponding to an expected value of 1 SNV per 1e4 bases.

Value

A vector of prior values with length given by the length of the regions GRanges object.

Note

Experimental code, subject to changes

Author(s)

mg14

Examples

## Make prior (not run)
#COSMIC <- readVcf("PATHTO/CosmicCodingMuts_v64_02042013_noLimit.vcf.gz", genome="GChr37")
#prior <- makePrior(COSMIC[info(COSMIC)$GENE=="TP53"], regions=GRanges(17, IRanges(7571720,7578811)))
#plot(prior[,1], type="h")

[Package deepSNV version 1.10.0 Index]