mcChunk {deepSNV}	R Documentation

Little helper function to split the count objects into a smaller digestible chunks and run function FUN on each subset

Description

Little helper function to split the count objects into a smaller digestible chunks and run function FUN on each subset

Usage

  mcChunk(FUN, X, split = 250, mc.cores = 1, ...)

Arguments

FUN
The function to call on each chunk

X
The object to be subsetted using [,i,]

split
The size of each chunk

mc.cores
The number of cores to use

...
Additional arguments passed to FUN

Value

The value of FUN

Note

Experimental code, subject to changes

Author(s)

mg14

[Package deepSNV version 1.10.0 Index]