normalize {deepSNV}	R Documentation

Normalize nucleotide counts.

Description

This functions performs a loess normalization of the nucleotide. This experimental feature can be used to compare experiments from different libraries or sequencing runs that may have differing noise characteristics.

Normalize nucleotide counts.

Normalize nucleotide counts.

Arguments

test
Either an deepSNV-class object or a named matrix with nucleotide counts.

control
Missing if test is an link{deepSNV-class} object, otherwise a matrix with nucleotide counts.

round
Logical. Should normalized counts be rounded to integers? Default=TRUE

...
Parameters passed to loess.

Value

A deepSNV-class object.

Note

This feature is somewhat experimental and the results should be treated with care. Sometimes it can be better to leave the data unnormalized and use a model with greater dispersion instead.

Author(s)

Moritz Gerstung

Examples

data(phiX, package = "deepSNV")
plot(phiX)
phiN <- normalize(phiX, round = TRUE)
plot(phiN)

[Package deepSNV version 1.10.0 Index]