p.combine {deepSNV}	R Documentation

Combine two p-values

Description

This function combines two P-values into a single one using a statistic defined by method. "fisher" uses the product of the two, in this case the logarithm of the product is χ^2_4 distributed. If the method = "max", the resulting P-value is \max\{P_1,P_2\}^2. For method = "average" the mean is used, yielding a P-value of 2 x^2 if x=(P_1+P_2)/2 < .5 and 1-2 x^2 otherwise.

Usage

  p.combine(p1, p2,
    method = c("fisher", "max", "average", "prod"))

Arguments

p1
P-value 1

p2
P-value 2

method
One of "fisher" (default), "max" or "average"

Value

p-values

Author(s)

Moritz Gerstung

Examples

p1 <- runif(1000)
p2 <- runif(1000)
hist(p1)
p.avg = p.combine(p1,p2, method="average")
hist(p.avg)
p.fish = p.combine(p1,p2, method="fisher")
hist(p.fish)
p.max = p.combine(p1,p2, method="max")
hist(p.max)
pairs(data.frame(p1,p2,p.fish,p.max,p.avg))

[Package deepSNV version 1.10.0 Index]