repeatMask {deepSNV}	R Documentation

Mask homopolymeric repeats.

Description

This function masks homopolymeric repeats longer than a given width. These are hot-spots of sequencing error and can confound the analysis.

Arguments

x
An object. Either a deepSNV-class object or a DNAString with the nucleotide sequence.

flank
Boolean. Indicates whether the sites adjacent to the repeat should also be masked.

w
Integer. The minimal length at which repeats should be masked. Default w=0.

Value

A boolean vector where TRUE indicates a non-homopolymeric region.

Author(s)

Moritz Gerstung

Examples

data(HIVmix)
which(repeatMask(HIVmix))

[Package deepSNV version 1.10.0 Index]